lsコマンドに対する操作(parse_options)をls.rbへ移管
