+++ 08.ls_object/ls.rb
@@ -1,5 +1,27 @@
 # frozen_string_literal: true
 
+require 'optparse'
 require_relative 'ls_main'
 
-FileList.new.format_files
+class LsCommand
+  def initialize
+    @options = parse_options
+  end
+
+  def parse_options
+    options = {}
+    OptionParser.new do |opts|
+      opts.on('-a') { options['a'] = true }
+      opts.on('-r') { options['r'] = true }
+      opts.on('-l') { options['l'] = true }
+    end.parse(ARGV)
+    options
+  end
+
+  def run
+    file_list = FileList.new(@options)
+    file_list.format_files
+  end
+end
+
+LsCommand.new.run
